qa/requirements.txt
python-dotenv
aiohttp
langchain
pandas
transformers
faiss-cpu